% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmchen.r
\name{pmchen}
\alias{pmchen}
\title{main function}
\usage{
pmchen(
  x,
  y,
  method = c("weighted_z", "mle_hetero", "mle_homo", "modified_t", "zcorr"),
  alternative = "two.sided"
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values(tumor sample)}

\item{y}{a (non-empty) numeric vector of data values(matched normal sample)}

\item{method}{a character string indicating which method is to be computed. One of the "weighted"(default),"mle_hetero","mle_hoto","zcorr",or "modified_t".}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
statistics p-value of several methods for handing partially matched sample (tumor and matched normal sample) test by using meta-analysis framework.
}
\description{
This is the function allows you to calculate p value
}
\examples{

set.seed(123)
n=20
x=rnorm(n) 
y=rnorm(n)
x[1:5]=NA
y[7:10]=NA
pmchen(x,y,method="weighted_z",alternative="two.sided")
pmchen(x,y,method="mle_hetero",alternative="two.sided")
pmchen(x,y,method="mle_homo",alternative="two.sided")
pmchen(x,y,method="zcorr",alternative="two.sided")
pmchen(x,y,method="modified_t",alternative="two.sided")
pmchen(x,y)
pmchen(x,y,method="modified_t",alternative="great")
}
